# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FeedbackHub API
  version: "1.0.0"
  description: |
    REST API for FeedbackHub, a self-hostable feedback & survey widget platform.
    Administrative endpoints require a Bearer access token. Widget-facing endpoints
    (under /api/v1/public and POST /api/v1/responses) are authenticated with a
    project public key instead of a user session.
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    projectKey:
      type: apiKey
      in: header
      name: X-Project-Key
...
paths:
...
  /widgets:
    get:
      summary: List widgets for a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: projectId, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: List of widgets }
    post:
      summary: Create a widget
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Widget created }
  /widgets/{id}:
    get:
      summary: Get a widget
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Widget }
    patch:
      summary: Update a widget
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Updated widget }
    delete:
      summary: Soft-delete a widget
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
  /widgets/{id}/publish:
    post:
      summary: Publish a widget (fires widget.published webhook)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Published widget }
  /widgets/{id}/unpublish:
    post:
      summary: Unpublish a widget
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Draft widget }
...